@@ -1,18 +1,20 @@
 // ignore_for_file: non_constant_identifier_names, dead_code
+import 'dart:async';
 import 'dart:io';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:responsive_framework/responsive_framework.dart';
 import 'package:videosdk/videosdk.dart';
 import 'package:videosdk_flutter_example/screens/conference-call/conference_meeting_screen.dart';
 import 'package:videosdk_flutter_example/utils/api.dart';
-import 'package:videosdk_flutter_example/widgets/common/joining_details/joining_details.dart';
-
+import 'package:videosdk_flutter_example/widgets/common/joining/join_options.dart';
 import '../../constants/colors.dart';
-import '../../utils/spacer.dart';
 import '../../utils/toast.dart';
+import '../../widgets/common/joining/join_view.dart';
 import '../one-to-one/one_to_one_meeting_screen.dart';
+import 'dropdowns_Web.dart';
+import 'selectAudioDevice.dart';
+import 'selectVideoDevice.dart';
 
 // Join Screen
 class JoinScreen extends StatefulWidget {
@@ -22,23 +24,43 @@ class JoinScreen extends StatefulWidget {
   State<JoinScreen> createState() => _JoinScreenState();
 }
 
-class _JoinScreenState extends State<JoinScreen> {
+class _JoinScreenState extends State<JoinScreen> with WidgetsBindingObserver {
   String _token = "";
 
   // Control Status
-  bool isMicOn = true;
-  bool isCameraOn = true;
+  bool isMicOn =
+      !kIsWeb && (Platform.isMacOS || Platform.isWindows) ? true : false;
+  bool isCameraOn =
+      !kIsWeb && (Platform.isMacOS || Platform.isWindows) ? true : false;
 
   CustomTrack? cameraTrack;
+  CustomTrack? microphoneTrack;
   RTCVideoRenderer? cameraRenderer;
 
   bool? isJoinMeetingSelected;
   bool? isCreateMeetingSelected;
 
+  bool? isCameraPermissionAllowed =
+      !kIsWeb && (Platform.isMacOS || Platform.isWindows) ? true : false;
+  bool? isMicrophonePermissionAllowed =
+      !kIsWeb && (Platform.isMacOS || Platform.isWindows) ? true : false;
+
+  VideoDeviceInfo? selectedVideoDevice;
+  AudioDeviceInfo? selectedAudioOutputDevice;
+  AudioDeviceInfo? selectedAudioInputDevice;
+  List<VideoDeviceInfo>? videoDevices;
+  List<AudioDeviceInfo>? audioDevices;
+  List<AudioDeviceInfo> audioInputDevices = [];
+  List<AudioDeviceInfo> audioOutputDevices = [];
+
+  late Function handler;
+
   @override
   void initState() {
-    initCameraPreview();
     super.initState();
+
+    WidgetsBinding.instance.addObserver(this);
+
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
       DeviceOrientation.portraitDown,
@@ -47,6 +69,152 @@ class _JoinScreenState extends State<JoinScreen> {
       final token = await fetchToken(context);
       setState(() => _token = token);
     });
+    checkandReqPermissions();
+    subscribe();
+  }
+
+  void updateselectedAudioOutputDevice(AudioDeviceInfo? device) {
+    if (device?.deviceId != selectedAudioOutputDevice?.deviceId) {
+      setState(() {
+        selectedAudioOutputDevice = device;
+      });
+      if (!kIsWeb) {
+        if (Platform.isAndroid || Platform.isIOS) {
+          disposeMicTrack();
+          initMic();
+        }
+      }
+    }
+  }
+
+  void updateselectedAudioInputDevice(AudioDeviceInfo? device) {
+    if (device?.deviceId != selectedAudioInputDevice?.deviceId) {
+      setState(() {
+        selectedAudioInputDevice = device;
+      });
+      disposeMicTrack();
+      initMic();
+    }
+  }
+
+  void updateSelectedVideoDevice(VideoDeviceInfo? device) {
+    if (device?.deviceId != selectedVideoDevice?.deviceId) {
+
+      disposeCameraPreview();
+      setState(() {
+        selectedVideoDevice = device;
+      });
+      
+      initCameraPreview();
+    }
+  }
+
+  void checkBluetoothPermissions() async {
+    try {
+      bool bluetoothPerm = await VideoSDK.checkBluetoothPermission();
+      if (bluetoothPerm != true) {
+        await VideoSDK.requestBluetoothPermission();
+      }
+    } catch (e) {}
+  }
+
+  void getDevices() async {
+    if (isCameraPermissionAllowed != null &&
+        isCameraPermissionAllowed == true) {
+      videoDevices = await VideoSDK.getVideoDevices();
+      setState(() {
+        selectedVideoDevice = videoDevices?.first;
+      });
+      initCameraPreview();
+    }
+    if (isMicrophonePermissionAllowed != null &&
+        isMicrophonePermissionAllowed == true) {
+      audioDevices = await VideoSDK.getAudioDevices();
+      if (!kIsWeb && !Platform.isMacOS && !Platform.isWindows) {
+        //Condition for android and ios devices
+        setState(() {
+          selectedAudioOutputDevice = audioDevices?.first;
+        });
+      } else {
+        audioInputDevices = [];
+        audioOutputDevices = [];
+        for (AudioDeviceInfo device in audioDevices!) {
+          if (device.kind == 'audioinput') {
+            audioInputDevices.add(device);
+          } else {
+            audioOutputDevices.add(device);
+          }
+        }
+        setState(() {
+          selectedAudioOutputDevice = audioOutputDevices.first;
+          selectedAudioInputDevice = audioInputDevices.first;
+        });
+        initMic();
+      }
+    }
+  }
+
+  void checkandReqPermissions([Permissions? perm]) async {
+    perm ??= Permissions.audio_video;
+    try {
+      if (kIsWeb || Platform.isAndroid || Platform.isIOS) {
+        Map<String, bool> permissions = await VideoSDK.checkPermissions();
+
+        if (perm == Permissions.audio || perm == Permissions.audio_video) {
+          if (permissions['audio'] != true) {
+            Map<String, bool> reqPermissions =
+                await VideoSDK.requestPermissions(Permissions.audio);
+            setState(() {
+              isMicrophonePermissionAllowed = reqPermissions['audio'];
+              isMicOn = reqPermissions['audio']!;
+            });
+          } else {
+            setState(() {
+              isMicrophonePermissionAllowed = true;
+              isMicOn = true;
+            });
+          }
+        }
+
+        if (perm == Permissions.video || perm == Permissions.audio_video) {
+          if (permissions['video'] != true) {
+            Map<String, bool> reqPermissions =
+                await VideoSDK.requestPermissions(Permissions.video);
+
+            setState(() => isCameraPermissionAllowed = reqPermissions['video']);
+          } else {
+            setState(() => isCameraPermissionAllowed = true);
+          }
+        }
+        if (!kIsWeb) {
+          if (Platform.isAndroid) {
+            checkBluetoothPermissions();
+          }
+        }
+      }
+      getDevices();
+    } catch (e) {}
+  }
+
+  void checkPermissions() async {
+    if (kIsWeb || Platform.isAndroid || Platform.isIOS) {
+      Map<String, bool> permissions = await VideoSDK.checkPermissions();
+      setState(() {
+        isMicrophonePermissionAllowed = permissions['audio'];
+        isCameraPermissionAllowed = permissions['video'];
+        isMicOn = permissions['audio']!;
+        isCameraOn = permissions['video']!;
+      });
+    }
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    switch (state) {
+      case AppLifecycleState.resumed:
+        checkPermissions();
+        break;
+    }
   }
 
   @override
@@ -59,244 +227,272 @@ class _JoinScreenState extends State<JoinScreen> {
   @override
   Widget build(BuildContext context) {
     final maxWidth = MediaQuery.of(context).size.width;
-    return WillPopScope(
-        onWillPop: _onWillPopScope,
+    final maxHeight = MediaQuery.of(context).size.height;
+    return PopScope(
+        canPop: false,
+        onPopInvoked: (didPop) async {
+          if (didPop) {
+            return;
+          }
+          _onWillPopScope();
+        },
         child: Scaffold(
+          appBar: !kIsWeb && (Platform.isAndroid || Platform.isIOS)
+              ? AppBar(
+                  flexibleSpace: Align(
+                    child: Padding(
+                      padding: EdgeInsets.fromLTRB(0, 40, 10, 0),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.end,
+                        children: [
+                          IconButton(
+                            icon: Icon(
+                              Icons.volume_up,
+                              size: 27,
+                              color: Colors.white,
+                            ),
+                            onPressed: () {
+                              showModalBottomSheet<void>(
+                                context: context,
+                                builder: (BuildContext context) {
+                                  return Container(
+                                      color: black750,
+                                      child: Padding(
+                                          padding: const EdgeInsets.symmetric(
+                                              vertical: 10),
+                                          child: SelectAudioDevice(
+                                            isMicrophonePermissionAllowed:
+                                                isMicrophonePermissionAllowed,
+                                            selectedAudioOutputDevice:
+                                                selectedAudioOutputDevice,
+                                            audioDevices: audioDevices,
+                                            onAudioDeviceSelected:
+                                                updateselectedAudioOutputDevice,
+                                          )));
+                                },
+                              );
+                            },
+                          ),
+                          IconButton(
+                            icon: Icon(
+                              Icons.camera_alt_rounded,
+                              size: 27,
+                              color: Colors.white,
+                            ),
+                            onPressed: () {
+                              showModalBottomSheet<void>(
+                                context: context,
+                                builder: (BuildContext context) {
+                                  return Container(
+                                    color: black750,
+                                    child: Padding(
+                                      padding: const EdgeInsets.symmetric(
+                                          vertical: 10),
+                                      child: SelectVideoDevice(
+                                        isCameraPermissionAllowed:
+                                            isCameraPermissionAllowed,
+                                        selectedVideoDevice:
+                                            selectedVideoDevice,
+                                        videoDevices: videoDevices,
+                                        onVideoDeviceSelected:
+                                            updateSelectedVideoDevice,
+                                      ),
+                                    ),
+                                  );
+                                },
+                              );
+                            },
+                          ),
+                        ],
+                      ),
+                    ),
+                  ),
+                  backgroundColor: black750,
+                  elevation: 0,
+                )
+              : null,
           backgroundColor: primaryColor,
           body: SafeArea(
             child: LayoutBuilder(
               builder:
                   (BuildContext context, BoxConstraints viewportConstraints) {
-                return SingleChildScrollView(
-                  child: ConstrainedBox(
-                    constraints: BoxConstraints(
-                        minHeight: viewportConstraints.maxHeight),
-                    child: IntrinsicHeight(
-                      child: Column(
-                        mainAxisAlignment:
-                            !kIsWeb && (Platform.isAndroid || Platform.isIOS)
-                                ? MainAxisAlignment.center
-                                : MainAxisAlignment.start,
-                        crossAxisAlignment: CrossAxisAlignment.stretch,
-                        children: [
-                          // Camera Preview
-                          Padding(
-                            padding: const EdgeInsets.symmetric(
-                                vertical: 100, horizontal: 36),
-                            child:
-                                SizedBox(
-                                  height: 300,
-                                  width: 200,
-                                  child: Stack(
-                                    alignment: Alignment.topCenter,
-                                    children: [
-                                      
-                                         AspectRatio(
-                                              aspectRatio: ResponsiveValue<double>(
-                                                  context,
-                                                  conditionalValues: [
-                                                    Condition.equals(
-                                                        name: MOBILE,
-                                                        value: 1 / 1.55),
-                                                    Condition.equals(
-                                                        name: TABLET,
-                                                        value: 16 / 10),
-                                                    Condition.largerThan(
-                                                        name: TABLET,
-                                                        value: 16 / 9),
-                                                  ]).value!,
-                                              child: cameraRenderer != null
-                                                  ? ClipRRect(
-                                                      borderRadius:
-                                                          BorderRadius.circular(12),
-                                                      child: RTCVideoView(
-                                                        cameraRenderer
-                                                            as RTCVideoRenderer,
-                                                        objectFit: RTCVideoViewObjectFit
-                                                            .RTCVideoViewObjectFitCover,
-                                                      ),
-                                                    )
-                                                  : Container(
-                                                      decoration: BoxDecoration(
-                                                          color: black800,
-                                                          borderRadius:
-                                                              BorderRadius.circular(
-                                                                  12)),
-                                                      child: const Center(
-                                                        child: Text(
-                                                          "Camera is turned off",
-                                                        ),
-                                                      ),
-                                                    ),
-                                            ),
-                                  Positioned(
-                                    bottom: 16,
-                                    // Meeting ActionBar
-                                    child: Center(
-                                      child: Row(
-                                        children: [
-                                          // Mic Action Button
-                                          ElevatedButton(
-                                            onPressed: () => setState(
-                                              () => isMicOn = !isMicOn,
-                                            ),
-                                            style: ElevatedButton.styleFrom(
-                                              shape: const CircleBorder(),
-                                              padding: EdgeInsets.all(
-                                                ResponsiveValue<double>(context,
-                                                    conditionalValues: [
-                                                      Condition.equals(
-                                                          name: MOBILE,
-                                                          value: 12),
-                                                      Condition.equals(
-                                                          name: TABLET,
-                                                          value: 15),
-                                                      Condition.equals(
-                                                          name: DESKTOP,
-                                                          value: 18),
-                                                    ]).value!,
-                                              ),
-                                              backgroundColor:
-                                                  isMicOn ? Colors.white : red,
-                                              foregroundColor: Colors.black,
-                                            ),
-                                            child: Icon(
-                                                isMicOn
-                                                    ? Icons.mic
-                                                    : Icons.mic_off,
-                                                color: isMicOn
-                                                    ? grey
-                                                    : Colors.white),
-                                          ),
-                                          ElevatedButton(
-                                            onPressed: () {
-                                              if (isCameraOn) {
-                                                disposeCameraPreview();
-                                              } else {
-                                                initCameraPreview();
-                                              }
-                                              setState(() =>
-                                                  isCameraOn = !isCameraOn);
-                                            },
-                                            style: ElevatedButton.styleFrom(
-                                              shape: const CircleBorder(),
-                                              padding: EdgeInsets.all(
-                                                ResponsiveValue<double>(context,
-                                                    conditionalValues: [
-                                                      Condition.equals(
-                                                          name: MOBILE,
-                                                          value: 12),
-                                                      Condition.equals(
-                                                          name: TABLET,
-                                                          value: 15),
-                                                      Condition.equals(
-                                                          name: DESKTOP,
-                                                          value: 18),
-                                                    ]).value!,
-                                              ),
-                                              backgroundColor: isCameraOn
-                                                  ? Colors.white
-                                                  : red,
-                                            ),
-                                            child: Icon(
-                                              isCameraOn
-                                                  ? Icons.videocam
-                                                  : Icons.videocam_off,
-                                              color: isCameraOn
-                                                  ? grey
-                                                  : Colors.white,
-                                            ),
-                                          ),
-                                        ],
-                                      ),
+                Widget _buildContent() {
+                  return IntrinsicHeight(
+                    child: kIsWeb || Platform.isWindows || Platform.isMacOS
+                        ? Container(
+                            margin: const EdgeInsets.fromLTRB(40, 150, 40, 150),
+                            child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceAround,
+                                children: [
+                                  SizedBox(
+                                    height: maxHeight,
+                                    width: maxWidth / 2,
+                                    child: Column(
+                                      children: [
+                                        JoinView(
+                                          cameraRenderer: cameraRenderer,
+                                          isMicOn: isMicOn,
+                                          isCameraOn: isCameraOn,
+                                          onMicToggle: () =>
+                                              isMicrophonePermissionAllowed !=
+                                                          null &&
+                                                      isMicrophonePermissionAllowed ==
+                                                          true
+                                                  ? {
+                                                      if (isMicOn)
+                                                        {disposeMicTrack()}
+                                                      else
+                                                        {initMic()},
+                                                      setState(
+                                                        () =>
+                                                            isMicOn = !isMicOn,
+                                                      )
+                                                    }
+                                                  : checkandReqPermissions(
+                                                      Permissions.audio),
+                                          onCameraToggle: () {
+                                            isCameraPermissionAllowed != null &&
+                                                    isCameraPermissionAllowed ==
+                                                        true
+                                                ? {
+                                                    if (isCameraOn)
+                                                      {disposeCameraPreview()}
+                                                    else
+                                                      {initCameraPreview()},
+                                                    setState(() => isCameraOn =
+                                                        !isCameraOn)
+                                                  }
+                                                : checkandReqPermissions(
+                                                    Permissions.video);
+                                          },
+                                        ),
+                                        kIsWeb ||
+                                                Platform.isMacOS ||
+                                                Platform.isWindows
+                                            ? Padding(
+                                                padding:
+                                                    const EdgeInsets.fromLTRB(
+                                                        0, 20, 0, 0),
+                                                child: DropdownsWebWidget(
+                                                  isCameraPermissionAllowed:
+                                                      isCameraPermissionAllowed,
+                                                  isMicrophonePermissionAllowed:
+                                                      isMicrophonePermissionAllowed,
+                                                  selectedAudioOutputDevice:
+                                                      selectedAudioOutputDevice,
+                                                  selectedAudioInputDevice:
+                                                      selectedAudioInputDevice,
+                                                  selectedVideoDevice:
+                                                      selectedVideoDevice,
+                                                  audioInputDevices:
+                                                      audioInputDevices,
+                                                  audioOutputDevices:
+                                                      audioOutputDevices,
+                                                  videoDevices: videoDevices,
+                                                  onAudioOutputDeviceSelected:
+                                                      updateselectedAudioOutputDevice,
+                                                  onAudioInputDeviceSelected:
+                                                      updateselectedAudioInputDevice,
+                                                  onVideoDeviceSelected:
+                                                      updateSelectedVideoDevice,
+                                                ),
+                                              )
+                                            : Container(),
+                                      ],
                                     ),
                                   ),
-                                ],
-                              ),
-                            ),
-                          ),
-                          Padding(
-                            padding: const EdgeInsets.only(bottom: 30),
-                            child: Column(
-                              children: [
-                                if (isJoinMeetingSelected == null &&
-                                    isCreateMeetingSelected == null)
-                                  MaterialButton(
-                                      minWidth: ResponsiveValue<double>(context,
-                                          conditionalValues: [
-                                            Condition.equals(
-                                                name: MOBILE,
-                                                value: maxWidth / 1.3),
-                                            Condition.equals(
-                                                name: TABLET,
-                                                value: maxWidth / 1.3),
-                                            Condition.equals(
-                                                name: DESKTOP, value: 600),
-                                          ]).value!,
-                                      height: 50,
-                                      shape: RoundedRectangleBorder(
-                                          borderRadius:
-                                              BorderRadius.circular(12)),
-                                      padding: const EdgeInsets.symmetric(
-                                          vertical: 16),
-                                      color: purple,
-                                      child: const Text("Create Meeting",
-                                          style: TextStyle(fontSize: 16)),
-                                      onPressed: () => {
-                                            setState(() => {
-                                                  isCreateMeetingSelected =
-                                                      true,
-                                                  isJoinMeetingSelected = true
-                                                })
-                                          }),
-                                const VerticalSpacer(16),
-                                if (isJoinMeetingSelected == null &&
-                                    isCreateMeetingSelected == null)
-                                  MaterialButton(
-                                      minWidth: ResponsiveValue<double>(context,
-                                          conditionalValues: [
-                                            Condition.equals(
-                                                name: MOBILE,
-                                                value: maxWidth / 1.3),
-                                            Condition.equals(
-                                                name: TABLET,
-                                                value: maxWidth / 1.3),
-                                            Condition.equals(
-                                                name: DESKTOP, value: 600),
-                                          ]).value!,
-                                      height: 50,
-                                      shape: RoundedRectangleBorder(
-                                          borderRadius:
-                                              BorderRadius.circular(12)),
-                                      padding: const EdgeInsets.symmetric(
-                                          vertical: 16),
-                                      color: black750,
-                                      child: const Text("Join Meeting",
-                                          style: TextStyle(fontSize: 16)),
-                                      onPressed: () => {
-                                            setState(() => {
-                                                  isCreateMeetingSelected =
-                                                      false,
-                                                  isJoinMeetingSelected = true
-                                                })
-                                          }),
-                                if (isJoinMeetingSelected != null &&
-                                    isCreateMeetingSelected != null)
-                                  JoiningDetails(
-                                    isCreateMeeting: isCreateMeetingSelected!,
-                                    onClickMeetingJoin: (meetingId, callType,
-                                            displayName) =>
-                                        _onClickMeetingJoin(
-                                            meetingId, callType, displayName),
-                                  ),
-                              ],
-                            ),
+                                  Container(
+                                    margin: EdgeInsets.only(top: maxWidth / 10),
+                                    child: JoinOptions(
+                                      isJoinMeetingSelected:
+                                          isJoinMeetingSelected,
+                                      isCreateMeetingSelected:
+                                          isCreateMeetingSelected,
+                                      maxWidth: maxWidth,
+                                      onOptionSelected: (isCreateMeeting) {
+                                        setState(() {
+                                          isCreateMeetingSelected =
+                                              isCreateMeeting;
+                                          isJoinMeetingSelected = true;
+                                        });
+                                      },
+                                      onClickMeetingJoin: _onClickMeetingJoin,
+                                    ),
+                                  )
+                                ]),
                           )
-                        ],
-                      ),
-                    ),
-                  ),
-                );
+                        : Column(
+                            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                            children: [
+                                SizedBox(
+                                  height: 400,
+                                  width: maxWidth,
+                                  child: JoinView(
+                                    cameraRenderer: cameraRenderer,
+                                    isMicOn: isMicOn,
+                                    isCameraOn: isCameraOn,
+                                    onMicToggle: () =>
+                                        isMicrophonePermissionAllowed != null &&
+                                                isMicrophonePermissionAllowed ==
+                                                    true
+                                            ? {
+                                                if (isMicOn)
+                                                  {disposeMicTrack()}
+                                                else
+                                                  {initMic()},
+                                                setState(
+                                                  () => isMicOn = !isMicOn,
+                                                )
+                                              }
+                                            : checkandReqPermissions(
+                                                Permissions.audio),
+                                    onCameraToggle: () {
+                                      isCameraPermissionAllowed != null &&
+                                              isCameraPermissionAllowed == true
+                                          ? {
+                                              if (isCameraOn)
+                                                {disposeCameraPreview()}
+                                              else
+                                                {initCameraPreview()},
+                                              setState(() =>
+                                                  isCameraOn = !isCameraOn)
+                                            }
+                                          : checkandReqPermissions(
+                                              Permissions.video);
+                                    },
+                                  ),
+                                ),
+                                JoinOptions(
+                                  isJoinMeetingSelected: isJoinMeetingSelected,
+                                  isCreateMeetingSelected:
+                                      isCreateMeetingSelected,
+                                  maxWidth: maxWidth,
+                                  onOptionSelected: (isCreateMeeting) {
+                                    setState(() {
+                                      isCreateMeetingSelected = isCreateMeeting;
+                                      isJoinMeetingSelected = true;
+                                    });
+                                  },
+                                  onClickMeetingJoin: _onClickMeetingJoin,
+                                )
+                              ]),
+                  );
+                }
+
+                return kIsWeb || Platform.isMacOS || Platform.isWindows
+                    ? ConstrainedBox(
+                        constraints: BoxConstraints(
+                            minHeight: viewportConstraints.maxHeight),
+                        child: _buildContent(),
+                      )
+                    : SingleChildScrollView(
+                        child: ConstrainedBox(
+                          constraints: BoxConstraints(
+                              minHeight: viewportConstraints.maxHeight),
+                          child: _buildContent(),
+                        ),
+                      );
               },
             ),
           ),
@@ -316,14 +512,34 @@ class _JoinScreenState extends State<JoinScreen> {
   }
 
   void initCameraPreview() async {
-    CustomTrack track = await VideoSDK.createCameraVideoTrack();
-    RTCVideoRenderer render = RTCVideoRenderer();
-    await render.initialize();
-    render.setSrcObject(stream: track.mediaStream, trackId: track.mediaStream.getVideoTracks().first.id);
-    setState(() {
-      cameraTrack = track;
-      cameraRenderer = render;
-    });
+    if (isCameraPermissionAllowed != null &&
+        isCameraPermissionAllowed == true ) {
+      CustomTrack? track = await VideoSDK.createCameraVideoTrack(
+          cameraId: selectedVideoDevice?.deviceId);
+      RTCVideoRenderer render = RTCVideoRenderer();
+      await render.initialize();
+      render.setSrcObject(
+          stream: track?.mediaStream,
+          trackId: track?.mediaStream.getVideoTracks().first.id);
+      setState(() {
+        cameraTrack = track;
+        cameraRenderer = render;
+        isCameraOn = true;
+      });
+    }
+  }
+
+  void initMic() async {
+    if (isMicrophonePermissionAllowed != null &&
+        isMicrophonePermissionAllowed == true) {
+      CustomTrack? track = await VideoSDK.createMicrophoneAudioTrack(
+          microphoneId: kIsWeb || Platform.isMacOS || Platform.isWindows
+              ? selectedAudioInputDevice?.deviceId
+              : selectedAudioOutputDevice?.deviceId);
+      setState(() {
+        microphoneTrack = track;
+      });
+    }
   }
 
   void disposeCameraPreview() {
@@ -331,6 +547,13 @@ class _JoinScreenState extends State<JoinScreen> {
     setState(() {
       cameraRenderer = null;
       cameraTrack = null;
+    });
+  }
+
+  void disposeMicTrack() {
+    microphoneTrack?.dispose();
+    setState(() {
+      microphoneTrack = null;
     });
   }
 
@@ -349,17 +572,25 @@ class _JoinScreenState extends State<JoinScreen> {
     try {
       var _meetingID = await createMeeting(_token);
       if (mounted) {
-        disposeCameraPreview();
+        setState(() {
+          cameraRenderer = null;
+        });
+        unsubscribe();
+
         if (callType == "GROUP") {
           Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => ConfereneceMeetingScreen(
+              builder: (context) => ConferenceMeetingScreen(
                 token: _token,
                 meetingId: _meetingID,
                 displayName: displayName,
                 micEnabled: isMicOn,
                 camEnabled: isCameraOn,
+                selectedAudioOutputDevice: selectedAudioOutputDevice,
+                selectedAudioInputDevice: selectedAudioInputDevice,
+                cameraTrack: cameraTrack,
+                micTrack: microphoneTrack,
               ),
             ),
           );
@@ -368,12 +599,15 @@ class _JoinScreenState extends State<JoinScreen> {
             context,
             MaterialPageRoute(
               builder: (context) => OneToOneMeetingScreen(
-                token: _token,
-                meetingId: _meetingID,
-                displayName: displayName,
-                micEnabled: isMicOn,
-                camEnabled: isCameraOn,
-              ),
+                  token: _token,
+                  meetingId: _meetingID,
+                  displayName: displayName,
+                  micEnabled: isMicOn,
+                  camEnabled: isCameraOn,
+                  selectedAudioOutputDevice: selectedAudioOutputDevice,
+                  selectedAudioInputDevice: selectedAudioInputDevice,
+                  cameraTrack: cameraTrack,
+                  micTrack: microphoneTrack),
             ),
           );
         }
@@ -392,18 +626,25 @@ class _JoinScreenState extends State<JoinScreen> {
     var validMeeting = await validateMeeting(_token, meetingId);
     if (validMeeting) {
       if (mounted) {
-        disposeCameraPreview();
+        setState(() {
+          cameraRenderer = null;
+        });
+        unsubscribe();
+
         if (callType == "GROUP") {
           Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => ConfereneceMeetingScreen(
-                token: _token,
-                meetingId: meetingId,
-                displayName: displayName,
-                micEnabled: isMicOn,
-                camEnabled: isCameraOn,
-              ),
+              builder: (context) => ConferenceMeetingScreen(
+                  token: _token,
+                  meetingId: meetingId,
+                  displayName: displayName,
+                  micEnabled: isMicOn,
+                  camEnabled: isCameraOn,
+                  selectedAudioOutputDevice: selectedAudioOutputDevice,
+                  selectedAudioInputDevice: selectedAudioInputDevice,
+                  cameraTrack: cameraTrack,
+                  micTrack: microphoneTrack),
             ),
           );
         } else {
@@ -411,12 +652,15 @@ class _JoinScreenState extends State<JoinScreen> {
             context,
             MaterialPageRoute(
               builder: (context) => OneToOneMeetingScreen(
-                token: _token,
-                meetingId: meetingId,
-                displayName: displayName,
-                micEnabled: isMicOn,
-                camEnabled: isCameraOn,
-              ),
+                  token: _token,
+                  meetingId: meetingId,
+                  displayName: displayName,
+                  micEnabled: isMicOn,
+                  camEnabled: isCameraOn,
+                  selectedAudioOutputDevice: selectedAudioOutputDevice,
+                  selectedAudioInputDevice: selectedAudioInputDevice,
+                  cameraTrack: cameraTrack,
+                  micTrack: microphoneTrack),
             ),
           );
         }
@@ -428,15 +672,26 @@ class _JoinScreenState extends State<JoinScreen> {
     }
   }
 
+  void subscribe() {
+    handler = (devices) {
+      getDevices();
+    };
+    VideoSDK.on(Events.deviceChanged, handler);
+  }
+
+  void unsubscribe() {
+    VideoSDK.off(Events.deviceChanged, handler);
+  }
+
   @override
   void dispose() {
+    unsubscribe();
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
       DeviceOrientation.portraitDown,
       DeviceOrientation.landscapeLeft,
       DeviceOrientation.landscapeRight,
     ]);
-    cameraTrack?.dispose();
     super.dispose();
   }
 }
